+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+- Added field for workflow wide variables (@BenchR267)
 - Added rerun field (@BenchR267)
 - Hotfix (changed stdin to stdout -.-) (@BenchR267)
 - Added test for Print (@BenchR267)